+++ lib/spectre.ex
@@ -34,7 +34,7 @@ defmodule Spectre do
   alias Spectre.Runtime
   alias Spectre.State
 
-  @version "0.1.5"
+  @version "0.1.6"
 
   @doc """
   Returns the running Spectre library version.